backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.10.3
fastapi-mail>=1.2.8
aiosmtplib>=2.0.2
openai>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pytz>=2023.3
python-dateutil>=2.8.2
gcsa>=2.1.0
websockets>=11.0.0
wsproto>=1.2.0
requests>=2.28.0
urllib3>=1.26.0
beautifulsoup4>=4.12.0
typing-extensions>=4.0.0
jinja2>=3.0.0
lxml>=4.9.0
